# medvision-cxr/frontend/public/openapi/medvision-cxr-openapi.yaml
openapi: 3.1.0
info:
  title: MedVision-CXR Backend
  version: 0.1.0
  summary: OpenAPI 3.1 specification for the MedVision-CXR backend
  description: MedVision-CXR is an AI-assisted chest X-ray triage system. This specification
    covers authentication, upload, AI-assisted risk assessment, Grad-CAM, doctor review,
    audit logs, model metadata, direct delete, health, and the governed deletion approval
    workflow.
paths:
...
  /api/v1/cxr/upload:
    post:
      tags:
      - cxr
      summary: Upload Cxr
      operationId: upload_cxr_api_v1_cxr_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_cxr_api_v1_cxr_upload_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/cxr/{image_id}/analyze:
    post:
      tags:
      - cxr
      summary: Analyze Cxr
      operationId: analyze_cxr_api_v1_cxr__image_id__analyze_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: string
          title: Image Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cxr/{image_id}/gradcam:
    post:
      tags:
      - cxr
      summary: Generate Gradcam
      operationId: generate_gradcam_api_v1_cxr__image_id__gradcam_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: string
          title: Image Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradCAMRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradCAMGenerateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cxr/results/{prediction_id}:
    get:
      tags:
      - cxr
      summary: Get Prediction Result
      operationId: get_prediction_result_api_v1_cxr_results__prediction_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: prediction_id
        in: path
        required: true
        schema:
          type: string
          title: Prediction Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionResultOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cxr/history:
    get:
      tags:
      - cxr
      summary: Get History
      operationId: get_history_api_v1_cxr_history_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/cxr/{image_id}/heatmap:
    get:
      tags:
      - cxr
      summary: Get Heatmap
      operationId: get_heatmap_api_v1_cxr__image_id__heatmap_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: string
          title: Image Id
      - name: label
        in: query
        required: true
        schema:
          type: string
          title: Label
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeatmapResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cxr/{image_id}:
    delete:
      tags:
      - cxr
      summary: Delete Cxr Image
      operationId: delete_cxr_image_api_v1_cxr__image_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: string
          title: Image Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteImageRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteImageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnalyzeRequest:
      properties:
        priority:
          type: string
          title: Priority
          default: normal
        requested_heatmaps:
          items:
            type: string
          type: array
          title: Requested Heatmaps
      type: object
      title: AnalyzeRequest
    AnalyzeResponse:
      properties:
        job_id:
          type: string
          format: uuid
          title: Job Id
        image_id:
          type: string
          format: uuid
          title: Image Id
        status:
          type: string
          title: Status
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - job_id
      - image_id
      - status
      title: AnalyzeResponse
...
    Body_upload_cxr_api_v1_cxr_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_cxr_api_v1_cxr_upload_post
    DeleteImageRequest:
      properties:
        delete_mode:
          type: string
          title: Delete Mode
          default: soft
        reason:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Reason
      type: object
      title: DeleteImageRequest
    DeleteImageResponse:
      properties:
        image_id:
          type: string
          format: uuid
          title: Image Id
        request_status:
          type: string
          title: Request Status
        delete_mode:
          type: string
          title: Delete Mode
        deleted_by:
          type: string
          format: uuid
          title: Deleted By
        completed_at:
          type: string
          format: date-time
          title: Completed At
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - image_id
      - request_status
      - delete_mode
      - deleted_by
      - completed_at
      title: DeleteImageResponse
...
    GradCAMGenerateResponse:
      properties:
        image_id:
          type: string
          format: uuid
          title: Image Id
        target_label:
          type: string
          title: Target Label
        heatmap_url:
          type: string
          title: Heatmap Url
        overlay_url:
          type: string
          title: Overlay Url
        notice:
          type: string
          title: Notice
          default: 模型在生成该风险提示时重点关注了以下区域。热力图仅用于辅助理解，不代表医学诊断依据。最终判断应由专业医生结合临床信息完成。
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - image_id
      - target_label
      - heatmap_url
      - overlay_url
      title: GradCAMGenerateResponse
    GradCAMRequest:
      properties:
        image_id:
          type: string
          format: uuid
          title: Image Id
        target_label:
          type: string
          title: Target Label
      type: object
      required:
      - image_id
      - target_label
      title: GradCAMRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    HeatmapResponse:
      properties:
        image_id:
          type: string
          format: uuid
          title: Image Id
        label:
          type: string
          title: Label
        heatmap_url:
          type: string
          title: Heatmap Url
        overlay_url:
          type: string
          title: Overlay Url
        notice:
          type: string
          title: Notice
          default: 热力图仅用于辅助理解，不代表医学诊断依据。
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - image_id
      - label
      - heatmap_url
      - overlay_url
      title: HeatmapResponse
    HistoryItemOut:
      properties:
        prediction_id:
          type: string
          format: uuid
          title: Prediction Id
        image_id:
          type: string
          format: uuid
          title: Image Id
        uploaded_at:
          type: string
          format: date-time
          title: Uploaded At
        overall_risk_level:
          type: string
          title: Overall Risk Level
        doctor_review_required:
          type: boolean
          title: Doctor Review Required
        uncertainty_flag:
          type: boolean
          title: Uncertainty Flag
        model_version:
          type: string
          title: Model Version
      type: object
      required:
      - prediction_id
      - image_id
      - uploaded_at
      - overall_risk_level
      - doctor_review_required
      - uncertainty_flag
      - model_version
      title: HistoryItemOut
    HistoryResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/HistoryItemOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - items
      - total
      title: HistoryResponse
    LabelProbabilityOut:
      properties:
        label:
          type: string
          title: Label
        risk_probability:
          type: number
          title: Risk Probability
        threshold:
          type: number
          title: Threshold
        risk_flag:
          type: boolean
          title: Risk Flag
      type: object
      required:
      - label
      - risk_probability
      - threshold
      - risk_flag
      title: LabelProbabilityOut
...
    PredictionResultOut:
      properties:
        prediction_id:
          type: string
          format: uuid
          title: Prediction Id
        image_id:
          type: string
          format: uuid
          title: Image Id
        model_version:
          type: string
          title: Model Version
        result_type:
          type: string
          title: Result Type
          default: AI-assisted risk assessment
        risk_assessment:
          $ref: '#/components/schemas/RiskAssessmentOut'
        triage_result:
          additionalProperties: true
          type: object
          title: Triage Result
        ai_assisted_findings:
          items:
            $ref: '#/components/schemas/LabelProbabilityOut'
          type: array
          title: Ai Assisted Findings
        doctor_review_suggestion:
          type: string
          title: Doctor Review Suggestion
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - prediction_id
      - image_id
      - model_version
      - risk_assessment
      - triage_result
      - ai_assisted_findings
      - doctor_review_suggestion
      title: PredictionResultOut
...
    RiskAssessmentOut:
      properties:
        overall_risk_level:
          type: string
          title: Overall Risk Level
        confidence_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence Score
        uncertainty_flag:
          type: boolean
          title: Uncertainty Flag
        doctor_review_required:
          type: boolean
          title: Doctor Review Required
      type: object
      required:
      - overall_risk_level
      - confidence_score
      - uncertainty_flag
      - doctor_review_required
      title: RiskAssessmentOut
...
    UploadResponse:
      properties:
        image_id:
          type: string
          format: uuid
          title: Image Id
        storage_key:
          type: string
          title: Storage Key
        quality_check:
          additionalProperties: true
          type: object
          title: Quality Check
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - image_id
      - storage_key
      - quality_check
      title: UploadResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
servers:
- url: http://localhost:8000
  description: Local development server